requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3
twilio==9.3.1
whisper==1.1.10
pillow==10.4.0
python-multipart==0.0.9
pytz==2024.1
dateparser==1.2.0
tenacity==8.5.0 